cast to uint8_t rather than int

+++ include/ccf/endpoint.h
@@ -525,7 +525,7 @@ struct formatter<ccf::endpoints::ForwardingRequired>
       default:
       {
         throw std::logic_error(fmt::format(
-          "Unhandled value for ForwardingRequired: {}", static_cast<int>(v)));
+          "Unhandled value for ForwardingRequired: {}", static_cast<uint8_t>(v)));
       }
     }
   }
